Crate mithril_aggregator_discovery

This crate provides mechanisms to discover aggregators in a Mithril network.

Modules

test

Test utilities.

Structs

AggregatorEndpoint

Representation of an aggregator endpoint

CapableAggregatorDiscoverer

An aggregator discoverer for specific capabilities.

HttpConfigAggregatorDiscoverer

An implementation of the AggregatorDiscoverer trait which discovers aggregators from remote networks configuration.

ShuffleAggregatorDiscoverer

A discoverer that returns a random set of aggregators

Enums

RequiredAggregatorCapabilities

Required capabilities for an aggregator.

Traits

AggregatorDiscoverer

An aggregator discoverer.