mithril_aggregator_discovery

Trait AggregatorDiscoverer

pub trait AggregatorDiscoverer: Sync + Send {
    // Required method
    fn get_available_aggregators<'life0, 'async_trait>(
        &'life0 self,
        network: MithrilNetwork,
    ) -> Pin<Box<dyn Future<Output = StdResult<Box<dyn Iterator<Item = AggregatorEndpoint>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

An aggregator discoverer.

Required Methods

fn get_available_aggregators<'life0, 'async_trait>( &'life0 self, network: MithrilNetwork, ) -> Pin<Box<dyn Future<Output = StdResult<Box<dyn Iterator<Item = AggregatorEndpoint>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get an iterator over a list of available aggregators in a Mithril network.

Note: there is no guarantee that the returned aggregators are sorted, complete or up-to-date.